+++ R/save-figure.R
@@ -138,60 +138,113 @@ prepare_figure <- function(figure,
 #' Output plots as files
 #'
 #' @inheritParams prepare_figure
-#' @param name Name of figure.
-#'    Used to name the output file(s) and embedded in the PDF document properties Title field.
-#' @param cropped Create a second output of the figure without margins or title. (Default: False)
-#' @param ext Extensions to be added to file names for the main figure and cropped figure. (Default: c("pdf", "png"))
+#' @param filename Name of file to create.
+#' @param cropped Name of second output file of the figure without margins or title.
 #' @param args List of arguments passed to `ggplot2::ggsave()` for the main figure.
 #' @param args_cropped List of arguments passed to `ggplot2::ggsave()` for the cropped figure.
+#' @param preview Preview the output in the RStudio Viewer pane. (Default: False)
 #' @param ... Other arguments passed to \link{prepare_figure}.
 #'
-#' @details
-#'
-#' # Using cairo_pdf
-#' Note that if you use a PDF device that does not have a `title`
-#' argument (such as cairo_pdf), then the `title` argument should be set to
-#' NULL, e.g. `args = list(device = cairo_pdf, title = NULL)`
 #'
 #' @export
 #'
 save_figure <- function(figure,
-                        name,
-                        cropped = FALSE,
-                        ext = c("pdf", "png"),
+                        filename,
+                        cropped = NULL,
                         args = NULL,
                         args_cropped = NULL,
+                        preview = FALSE,
                         ...){
 
   # Prepare figure
   figure <- prepare_figure(figure, ...)
 
   # Save to PDF file
-  figargs <- list(filename = paste0(name, ".", ext[[1]]),
+  figargs <- list(filename = filename,
                   plot     = figure$page,
                   width    = attr(figure$page, "width"),
                   height   = attr(figure$page, "height"),
                   units    = "mm",
                   bg       = "transparent")
-  if (ext[[1]] == "pdf"){figargs$title = name}
   if(!is.null(args)){figargs <- utils::modifyList(figargs, args)}
+
+  if (preview) {
+    do.call("ggpreview", figargs)
+    return(invisible(filename))
+  }
+
   do.call("ggsave", figargs)
 
   ## Save cropped figure to PNG file
-  if (cropped){
-    figargs <- list(paste0(name, "_cropped.", ext[[2]]),
+  if (!is.null(cropped)){
+    if (!is.character(cropped)) {rlang::abort("cropped should be a file name.")}
+    figargs <- list(filename = cropped,
                     plot   = figure$figure,
                     width  = attr(figure$figure, "width"),
                     height = attr(figure$figure, "width"),
                     units = "mm",
                     bg = "transparent")
-    if (ext[[2]] == "pdf"){figargs$title = name}
     if(!is.null(args_cropped)){figargs <- utils::modifyList(figargs, args_cropped)}
     do.call("ggsave", figargs)
   }
 
-  invisible(name)
+  return(invisible(filename))
 }
 
 
 
+
+
+
+#' Create a plot preview and display it in the Viewer pane.
+#'
+#' This function saves a ggplot2 plot to a temporary PNG file and then embeds it in an HTML
+#' page, which is opened in the Viewer pane.
+#'
+#' @details # Device
+#' The plot is saved using `ggsave` with the `png` device, regardless of what is specified in
+#' the call, so any arguments not used by `ggsave` or `png` are ignored.
+#'
+#' @param ... Arguments passed to `ggsave` and the PNG device function.
+#'
+#' @export
+#'
+ggpreview <- function(...) {
+  ## create temporary files
+  temp_img <- tempfile()
+  temp_html <- tempfile(fileext = ".html")
+
+  ## save using png device
+  call <- as.list(match.call())[-1]
+  call <- utils::modifyList(call,
+                            list(filename = temp_img,
+                                 device = "png"))
+  call <- call[names(call) %in% c(names(formals(ggsave)),
+                                  names(formals(grDevices::png)))]
+  do.call("ggsave", call)
+
+  ## create html file
+  html <- c("<!DOCTYPE html>",
+            "<html>",
+            "<title>Plot preview</title>",
+            "</head>",
+            "<body>",
+            glue::glue("<img src = '{xfun::base64_uri(temp_img, type = 'image/png')}' ",
+                       "style='margin: auto; border: 1px lightgrey solid; ",
+                       "display: block; ",
+                       "background-color: white;",
+                       "box-shadow: rgba(100, 100, 111, 0.75) 0px 7px 29px 0px; ",
+                       "max-width: 94%; max-height:94vh; height: auto;'"),
+            "</body>",
+            "</html>")
+  con <- file(temp_html, open = "w", encoding = "UTF-8")
+  writeLines(
+    html,
+    con = con,
+    sep = "\n")
+  close(con)
+
+  ## show html file in Viewer pane
+  viewer <- getOption("viewer", default = function(url){})
+  viewer(temp_html)
+}