+++ NEWS.md
@@ -1,3 +1,11 @@
+# ckbplotr 0.10.0
+
+* Added ggpreview() and preview argument in save_figure().
+* forest_plot() can handle lists for xlim (axis limits).
+* Added arguments to forest_plot() to control column spacing.
+* Improved generated code.
+* Fixes.
+
 # ckbplotr 0.9.2
 
 * Fix incorrect argument name in shape_plot().